Linux 2.2.19pre4
o	Fixed duplicate info on the microcode driver	(Daniel Rogers)
o	Update watchdog structs for nice user export	(Eric Brower)
o	Update Documentation/devices.txt		(H Peter Anvin)
o	Tweak sched.h to handle limit in Sparc		(Andrea Arcangeli)
	'make check_asm'
o	Move isdn pci definitions into pci.h		(Kai Germaschewski)
o	Tidy init data/static vars in the isdn code	(Kai Germaschewski)
o	Fix abuse of int for bitops in isdn		(Kai Germaschewski)
o	Use named initializers on the AVM B1		(Kai Germaschewski)
o	Switch capi message length to unsigned		(Kai Germaschewski)
o	ISDN updates					(Kai Germaschewski)
o	Update microcode code to check features right	(Tigran Aivazian)
	in 2.2
o	E820 handling fixup 				(Andrea Arcangeli)
o	Fix ne2k-pci driver build bug 			(J.A. Magallon)
o	DC390 driver updates				(Kurt Garloff)
o	Handle thinkpad E820 edx overwriting		(Marc Joosen)
o	Update the osst driver to 0.8.6.1		(Kurt Garloff,
							 Willem Riede)
o	Init the cmpci if compiled in		(Raúl Núñez de Arenas Coronado)
o	ATP870U SCSI updates to fix disconnect bug	(Wittman Li)
o	Clean up the usbdevfs backport			(Dan Streetman)
o	Fix ATI rage makefiles				(Brad Douglas)

+++ drivers/isdn/hisax/hisax.h
@@ -1,4 +1,4 @@
-/* $Id: hisax.h,v 2.49 2000/09/08 22:55:59 werner Exp $
+/* $Id: hisax.h,v 2.52.6.1 2000/12/06 16:59:19 kai Exp $
  *
  *   Basic declarations, defines and prototypes
  *
@@ -201,7 +201,7 @@ struct Layer1 {
 	void *hardware;
 	struct BCState *bcs;
 	struct PStack **stlistp;
-	int Flags;
+	long Flags;
 	struct FsmInst l1m;
 	struct FsmTimer	timer;
 	void (*l1l2) (struct PStack *, int, void *);
@@ -240,7 +240,7 @@ struct Layer2 {
 	int tei;
 	int sap;
 	int maxlen;
-	unsigned int flag;
+	unsigned long flag;
 	unsigned int vs, va, vr;
 	int rc;
 	unsigned int window;
@@ -365,7 +365,7 @@ struct w6692B_hw {
 };
 
 struct isar_reg {
-	unsigned int Flags;
+	unsigned long Flags;
 	volatile u_char bstat;
 	volatile u_char iis;
 	volatile u_char cmsb;
@@ -482,7 +482,7 @@ struct amd7930_hw {
 struct BCState {
 	int channel;
 	int mode;
-	int Flag;
+	long Flag; /* long req'd for set_bit --RR */
 	struct IsdnCardState *cs;
 	int tx_cnt;		/* B-Channel transmit counter */
 	struct sk_buff *tx_skb; /* B-Channel transmit Buffer */
@@ -527,7 +527,8 @@ struct Channel {
 	int data_open;
 	struct l3_process *proc;
 	setup_parm setup;	/* from isdnif.h numbers and Serviceindicator */
-	int Flags;		/* for remembering action done in l4 */
+	long Flags;		/* for remembering action done in l4 */
+				/* long req'd for set_bit --RR */
 	int leased;
 };
 
@@ -864,7 +865,7 @@ struct IsdnCardState {
 	int protocol;
 	unsigned int irq;
 	unsigned long irq_flags;
-	int HW_Flags;
+	long HW_Flags;
 	int *busy_flag;
         int chanlimit; /* limited number of B-chans to use */
         int logecho; /* log echo if supported by card */
@@ -931,7 +932,7 @@ struct IsdnCardState {
 	int rcvidx;
 	struct sk_buff *tx_skb;
 	int tx_cnt;
-	int event;
+	long event;
 	struct tq_struct tqueue;
 	struct timer_list dbusytimer;
 #ifdef ERROR_STATISTIC
